requirements.txt
absl-py==1.0.0
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.3.4
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.8
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.9
entrypoints==0.3
filelock==3.3.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing_extensions==4.0.0
urllib3==1.26.5
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.3
zipp==3.6.0